+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+1.2.4
+-----
+#. Validate member profile image strictly.
+#. Allow = in username.
+#. Friendly error message when attempting to use the same slug for overlapping sites.
+
 1.2.3.1
 -------
 #. Use new version of `django-ckeditor` with prettier toolbars.